completed separating forms into separate components
have to do validation
